creacion de clase GestionServicios donde se almacenan todos los metodos funcionales
